Patch the correct .mk when on linux.
Previously, we assumed that we'd be on a darwin system, but the .mk
file referenced during the buildPhase will depend on which operating
system we are on.

Note that assuming (!darwin => linux) is not actually safe, since it
is possible we may actually be running on some other unix platform.